Replace UnwrappedResults with more useful option
It probably isn't useful to have UnwrappedResults because
when you want it, you want it on a method chain, but it requires
the caller to check in advance if the value exists or not.

Intead we can proide a more useful method which maps to a common
case in Go: an empty T and an error.